Bring back browsermob-proxy for http1

+++ python/benchmark.py
@@ -1,6 +1,7 @@
 from selenium import webdriver
 from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
 import chromedriver_binary
+from browsermobproxy import Server
 
 import os
 import time
@@ -9,26 +10,17 @@ import json
 import psutil
 import shutil
 
-# Image sources.
-TEST_IMAGE_GCS_HTTP2 = {
-    "name": "gcs_http2_vanderbilt",
-    "url": "https://viv-demo.storage.googleapis.com/Vanderbilt-Spraggins-Kidney-MxIF.ome.tif",
-    "use_http2": True,
-}
-
-TEST_IMAGE_GCS_HTTP1 = {
-    "name": "gcs_http2_vanderbilt",
-    "url": "https://viv-demo.storage.googleapis.com/Vanderbilt-Spraggins-Kidney-MxIF.ome.tif",
-    "use_http2": False,
-}
-
-TEST_IMAGES = [TEST_IMAGE_GCS_HTTP1, TEST_IMAGE_GCS_HTTP2]
-
 # Top level directory for results.
 RESULTS_DIR = "../results/"
+# Change to true in order to use BMP
+USE_BMP_HTTP1 = False
+CAPTURE_HAR_WITH_BMP = False
 
 
-def end_daemon_chrome_processes():
+def make_har_json_name(url, is_http2):
+    return f"har_file_{url.split('/')[-1]}_{'http1' if not is_http2 else 'http2'}.json"
+
+def end_daemon_processes():
 
     for proc in psutil.process_iter():
         cmd = []
@@ -37,35 +29,46 @@ def end_daemon_chrome_processes():
             cmd = proc.cmdline()
         except:
             continue
-        if any(["chrome" in arg or "Chrome" in arg for arg in cmd]):
+        if any(["chrome" in arg or "Chrome" in arg or "browsermob-proxy" in arg for arg in cmd]):
             proc.kill()
+    time.sleep(1)
 
+def start_proxy_server():
+    # Unzipped library from `run-benchmark.sh`.
+    server = Server("./browsermob-proxy-2.1.4/bin/browsermob-proxy")
+    server.start()
+    proxy = server.create_proxy()
+    return (server, proxy)
 
-def start_chrome(use_http2):
+def start_chrome(is_http2, proxy_url, use_bmp_http1, capture_har_with_bmp):
 
     # Set up browser to launch.
     options = webdriver.ChromeOptions()
     options.add_argument(
-        f"--disable-infobars --disable-extensions --window-size=1900,900 {'--disable-http2' if not use_http2 else ''}"
+        "--disable-infobars --disable-extensions"
     )
+    if not is_http2 and use_bmp_http1:
+        options.add_argument(f"--proxy-server={proxy_url}")
+        options.add_argument("--ignore-certificate-errors")
+    if not capture_har_with_bmp:
+        options.add_extension("../js/har_extension.crx")
     options.add_argument("--auto-open-devtools-for-tabs")
-    options.add_extension("../js/har_extension.crx")
     prefs = {"download.default_directory": os.path.abspath(RESULTS_DIR)}
     options.add_experimental_option("prefs", prefs)
     driver = webdriver.Chrome(options=options)
+    driver.set_window_size(1920,1080)
     return driver
 
 
 def log_results(image):
 
     url = image["url"]
-    name = image["name"]
-    use_http2 = image["use_http2"]
+    is_http2 = image["is_http2"]
+    capture_har_with_bmp = image["capture_har_with_bmp"]
 
     # Get the har file, named after the file name fetched and the protocol (see background.js)
-    har_json_file = (
-        f"har_file_{url.split('/')[-1]}_{'http1' if not use_http2 else 'http2'}.json"
-    )
+    har_json_file = make_har_json_name(url, is_http2)
+
     with open(RESULTS_DIR + har_json_file, "r") as f:
         har_log = json.loads(f.read())["log"]
 
@@ -76,7 +79,7 @@ def log_results(image):
 
     # Get the timings in detail
     image_timings = [entry["timings"] for entry in image_entries]
-    print(f"***** TIMINGS FOR {url} {'http1' if not use_http2 else 'http2'} *******")
+    print(f"***** TIMINGS FOR {url} {'http1' if not is_http2 else 'http2'} {'captured with BMP' if capture_har_with_bmp else ''} *******")
 
     # Print each kind of timing and the total/average.
     print("Total time spent waiting: ", sum([entry["time"] for entry in image_entries]))
@@ -87,7 +90,7 @@ def log_results(image):
     for val in image_timings[0].keys():
         print(
             f"Time spent {val}: ",
-            sum([timing[val] for timing in image_timings if timing[val] > 0]),
+            sum([timing[val] for timing in image_timings if type(timing[val]) != str and timing[val] > 0]),
         )
     print("*******************************************")
 
@@ -95,33 +98,52 @@ def log_results(image):
 def run_test(image):
 
     url = image["url"]
-    name = image["name"]
-    use_http2 = image["use_http2"]
+    is_http2 = image["is_http2"]
+    use_bmp_http1 = image["use_bmp_http1"]
+    capture_har_with_bmp = image["capture_har_with_bmp"]
 
-    driver = start_chrome(use_http2)
+    (server, proxy) = start_proxy_server()
+    proxy_url = urlparse(proxy.proxy).path
+    driver = start_chrome(is_http2, proxy_url, use_bmp_http1, capture_har_with_bmp)
+    proxy.new_har(
+        f"localhost:9000/?image_url={url}",
+        options={"captureHeaders": True, "captureContent": True},
+    )
     driver.get(f"localhost:9000/?image_url={url}")
 
     # This should be more than enough for the test to complete and the har to download.
     # It is 10 seconds longer than when the har download starts.
     time.sleep(50)
 
+    # After finishing, download results from proxy of har.
+    if not is_http2 and use_bmp_http1 and capture_har_with_bmp:
+        os.makedirs(RESULTS_DIR, exist_ok=True)
+        with open(f"{RESULTS_DIR}/{make_har_json_name(url, is_http2)}", "w+") as f:
+            f.write(json.dumps(proxy.har, ensure_ascii=False))
+
     # Clean up.
+    server.stop()
     driver.quit()
 
 
 def run_benchmark():
 
     # Clear out previous results.
-    shutil.rmtree(RESULTS_DIR)
+    if os.path.exists(RESULTS_DIR) and os.path.isdir(RESULTS_DIR):
+        shutil.rmtree(RESULTS_DIR)
 
-    for image in TEST_IMAGES:
-        end_daemon_chrome_processes()
+    # Open configuration file.
+    with open('../image_test_files.json', 'r') as f:
+        images = json.loads(f.read())
+
+    for image in images:
+        end_daemon_processes()
 
         # Run tests and log results.
         run_test(image)
         log_results(image)
 
-        end_daemon_chrome_processes()
+        end_daemon_processes()
 
 
 if __name__ == "__main__":
